Various tweaks
- Move test functions to non-dev dependencies as are about to be used
  outside of dev
- Fix link to the XSD schema for PHPUnit to always have the up to date
  version
- Leverage nikic/iter instead of custom functions
- Dump the autoloader when the `composer.json` is changed via Makefile
  rules
- Leverage Assert for some checks when retrieving the compactors
